Frameworks: Added null checks for the view when SoftKeypad is launched and a list item is clicked
When the focus is on 2nd or 3rd MsgListItem and then SoftKeypad is
launched, the first position of msglistitem changes as the list scrolls
down.If we select the item which is not in the view group, getChildAt
returns null (as the index has gone outOfBounds).
So added a null check for view to fix the problem.

CRs-Fixed: 181850,181506