frameworks/base: Rename GLES2 libraries. Remove unneccessary code from SurfaceFlinger.cpp. Remove unneccessary files.
Rename GLES2 libraries to match Khronos recommendations:

libgles20.so -> libhGLESv2.so
libgles11.so -> libhGLESv1_CM.so
libegl13.so -> libhEGL.so

Clean up the GLES2 patch - libraries are now closed when we are done to avoid memory leaks.
Remove a questionable change from SurfaceFlinger.cpp that affects operation of the GLES2 patch.
Remove unneccessary files from the GLES and EGL wrapper libraries.