frameworks/base: Fix "-" character missing in KeyListerner
The "-" character wasn't mapped to any keycodes of the keyboard.
None of the KEYCODES in the sparsearray of KeyListener includes this character.
So, the "-" character is assigned to KEYCODE_1 as this keycode contains all the other special characters like .,/@ etc..
The corresponding changes were made to the file MultiTapKeyListener.java.
With this, the "-" character can be obtained by pressing 1 multiple times.

CRs-Fixed: 180253