SearchDialog: Cancel the filter Operation when the dialog is cancelled
Whenever the search dialog is invoked, a new filter thread starts and filters the database contents.
This operation is asynchronous and when the filtering operation is complete,
the thread exits. If the search dialog is invoked and immediately
cancelled, the filter thread continuous to run in the background until
it complete filtering. With monkey, as the events are generated very fast, more and more
filter threads are created and this finally results to ANR as the CPU
utilization increases. Thus cancelling the filtering operation when the
search dialog is cancelled.

CRs-Fixed: 184109