Contacts: Fix null pointer exception in cursor handling.
While in CallScreen,the photo of the contact is displayed by fetching it
either from the database or by displaying the default photo.
A Cursor is obtained by querying the contacts content provider for
photoUri using a content resolver.  Just after this query happens,
the acore process dies and the content providers get dead (making the
cursor null).  Before the acore process restarts and the content
providers are published again, this cursor is accessed and hence a null
pointer exception occurs.

CRs-Fixed: 186045