Frameworks/base: Fixing the incorrect selection of  key character map file for emulator.
On emulator, in some applications like Launcher, dialer, contacts etc., key events
related to music application do not get passed to the music app. Android was trying
to open key character map file, which is hardcoded as surf_keypad.kcm. As this file
will not be present in emulators, framework was getting rebooted because of null pointer
exception. Fixed it now by adding  more conditions to open proper file name based on device
in use.