Frameworks: Properly redraw progress bar on video resume
In case of wide video play back, when the Video app is suspended the surface
view that contains the video is destroyed, and when the video is resumed the
surface will be created again using the stretched video size which doesn't
cover the entire screen. This leaves a portion of the screen to not be redrawn during
playback, so when the progress bar shows/fades there's a portion of it left in
the bottom of the screen..

By resetting the size to 0 when the video is suspended, the surface will always
be drawn covering the entire screen when it resumes and the entire progress bar
portion in the screen will be redrawn.

CRs-Fixed: 188607