RIL: Modify OEM RIL request raw to return error code and response data
sendOemRilRequestRaw method does not return error code. Caller does not
know if the request was executed successfully or not. Modify the method
to return an error code. The method used to return a byte array which
is the response to the request. That is still needed by the caller. So
pass a byte array as a parameter to the method which upon return will
contain the response.