RIL: Fix race conditions during radio technology change
When Radio Technology changes from CDMA to GSM, the old Phone
instance is switched off and the new one is switched on without
waiting for the switch off to complete. There was a race condition
in switching from CDMA to GSM wherein, the CDMA phone switch off
caused a SIM to switch off after GSM was switched on. It was also
analyzed that the switch off command need not be sent from Android
for radio technology switch.
Hence removing the switch off and switch on solves the race condition
and also makes radio technology switch faster.

CRs-Fixed: 186087

Change depends on Gerrit Change #4813