WebView: Fix nullPointerException in WebView while clicking on a listbox
When a listbox in a webview is opened and immediately a new Browser
window is opened by pressing MENU+w, the mWebViewCore associated with
the listbox window becomes null. If an item in the list box is clicked,
it throws null pointer exception on mWebViewCore. So adding a null check
for the mWebViewCore to guard this race condition

CRs-Fixed: 191529