frameworks/base: Fix side effect of CR 188607
Apart from reseting video resolution in order to recreate entire surface, we
need to reset SurfaceView stored video dimensions in order to remeasure the
dimensions and draw correctly instead of stretching.

On Cupcake: When resuming a Video there's some shifting of the position of the video
going on as there are some changes in UI happening in background, if progress bar is
drawn on screen when this is happening, there's a part of it that is left on screen.
To fix this the view has to be redrawn when this shifts happen.

CRs-Fixed: 188607