frameworks/base: Remove the cached variable so that Exif Tags are parsed for each image individually.
In the Exif Interface, a singleton object is maintained to access the native
Exif library code. When the getattributes() is called for multiple images, it
incorrectly returns the previously stored exif information of previous image.
Thus all the images in the gallery have same exif taginformation. Parsing the
exif header individually resolves this issue.

CRs-Fixed: 193479