Wallpaper: Creation of wallpaper file's parent directory, while setting wallpaper
While running the monkey on Settings package, at some point of time, the
wallpaper file gets deleted, along with it's parent directory
[/data/data/com.android.settings/files]. Hence, when the launcher tries
to set the wallpaper again, it fails with a "file not found" exception
as it is not able to create the file. Hence, added a piece of code to
create the "files" directory, if it does not exist.

CRs-Fixed: 210968