telephony: Fixed erroneous roaming indicator
When we receive response to GET_REGISTRATION_STATE message from RIL
it contains few fields set to NULL. Code used to catch exception
as soon as it receives such message failing to parse the rest of the
valid data. This patches makes code ignore all fields that are NULL
while letting it parse non-NULL data.

CRs-fixed: 211291