framworks/base MediaPlayer: Handle suspend in paused state while streaming
Handling case where system goes to suspend while video is paused, and user
clicks play button as soon as progress bar comes up on resume.

PV state when this happens is STARTED as it's in auto pause state, because of
this isPlaying PV is returning isPlaying as true, though the correct state is
PREPARED so isPlaying should be false.

CRs-Fixed: 212037