Bluetooth: Modified FTP to process FTP transfer complete signal
Modified JNI and Service middleware to process FTP server transfer
complete signal. Prior to this change the signal was ignored and
files received via FTP server were not scanned by the upper layers
to import the files into the rspective applications. This change
propogates the signal up to the upper layers so the files can
be scanned.

CRs-fixed: 217872