frameworks/base: Provide APIs for storing allocations.
Current implementation does not allow applications to re-use already
created allocations. This feature is useful in case of orientation
change, where we might like to re-use the memory created earlier.
For example, Launcher2 changes orientation upon key board event,
it will re-create the icons if it were not allowed to keep them.
This change facilitates this requirement.