Telephony: Add subaddress to SMS fingerprint calculation.
As per SMS specification in 3GPP2 C.S0015-B, section 4.3.1.6, if the
Subaddress is included in a CDMA SMS message, it needs to be used for
duplicate detection. Subaddress, which is an optional field was omitted
while computing the SMS fingerprint. Hence it was never being used in
duplicate detection if it was included in the SMS. Add subaddress to the
SMS fingerprint.

Change-Id: I2206c130ac7d55a209d64c1baa98c4ebb4eb6e16
CRs-Fixed: 235885