GPS:Add "address" field to GPS NI notification
In some test cases, the LCS External Address needs to be displayed.The
address is added to the notification in addition to the existing client
name and requestor id. Dependency: need a change in libloc to provide
the
address in "extras" as Address = (the addr).Related CR for Eclair:223458
Also, the GSM decoder in GpsNetInitiatedHandler.java had a bug  which
caused the last letter to be dropped in some cases.

Change-Id: I6be093c5acd8f198a11de2b9bd563dc7366a59dc
CRs-Fixed: 247980