telephony: Storing voice mail count in phone for GSM
Voice mail icon was not displayed after phone reset
Voice mail indication for GSM is stored in SIM and if the
SIM does not support file id EF_MWIS which is an optional
field then the voice mail count was lost.

Added support to store voice mail count in phone memory
if SIM does not support this file id.

CRs-Fixed: 246305
Change-Id: I5f55cfb8abbd8f5e40b14d61e9604768fdfe0fb9