Telephony: Implement synchronous sending of sms messages
The code reads persist.radio.sms_sync_sending system property
to determine if synchronous sending algorith shoud be used.
When it is set to true SmsDispacher will wait for a message to be
sent before sending request for subsequent message while queueing all
messages that are requested to be sent. When this propery is not set
or is set to false the SmsDispatcher behaves as it used to by sending
all messages as soon as it receives the request to send them.

Change-Id: I8e157fdfbe0cc77be25eef73ff1359e829227316
CRs-Fixed: 219656