libstagefright: Fix for crash in processing extradata for flash.
ProcessExtradataOfBuffer function does not work
when OMXCodec is running in different process from
Mediaserver, since it is trying to access BufferHeader/pmem
pointer which are valid only in Mediaserver process.
Presently, there is no requirement of this function
in such cases either. So, avoiding call to this
function in such a case.

Change-Id: Ie05e5f4fa6ca92594335ee9fed5be1e205900885