media/libstagefright: Update buffering status based on maximum cached value instead of minimum
With the earlier code
1. Audio and Video streams were cached.
2. Total duration was available with the client.
3. In clips where this issue was seen, Audio and Video stream duration was not matching.
One of the streams was shorter  earlier than other.
4. client calculates the buffering status as per the below logic
Percentage = (Current Play  position + Minimum of Audio and  Video Cached duration)/Clip duration
In the clips where Audio and Video duration doesnt match, the above calculation
shows the buffering status based on the stream with lesser cached duration, assuming that both streams
are required to play the clip.
The above logic of percentage calculation is changed  and instead of minimum cache duration,
the maximum cache dutration of the 2 streams is used.

With this change the buffering status is shown properly.

Change-Id: I7e3bbf66f6a1dee2c1c7c5e06312b17ab4280b38
CRs-fixed: 257274