frameworks/base: Ignore unknown key events in VideoView
On 8660 FFA, pressing the back button generates 4 keyevents. The second and third
key events are unknown. Currently, the UNKNOWN event is handled
incorrectly by showing the media controls, preventing the BACK keyevent
from being handled. The UNKNOWN event is ignored to allow the superclass
to handle it. This change adds KEYCODE_UNKNOWN to the list of key events
that are not handled by VideoView

Change-Id: I0ef6333ecb267a5fab68ad72838a762ee7e5266d