libstagefright: Fix seek issues for short audio track clips
If seek was performed past the length of the audio track, the video
would freeze to wait for the fallback timer to catch up to the video.
Now we adjust the delta between fallback timer and video time so that
we no longer need to wait for the timer to catch up.

Change-Id: I7623ecb9376f383d76aa7974489b2efe9f93220c