mediaserver: Wait for services to die before attempting to start them
When restarting services immediately after a crash addService() can fail
because the service manager might not have noticed that the service had
crashed and updated its list of active services.  So the second
addService() will fail, since the service manager thinks that the
service is still running. This results in the service not restarting.
This change waits to add the service until the service manager knows
that the previous instance of the service is dead.

Change-Id: I3dcbeddf01f988eb6cddd6b9c88e5cb85211989c
CRs-Fixed: 257471