Telephony: Check for exception before processing message from RIL.
If modem comes up in GSM based network modes, GET_CDMA_SUBSCRIPTION_SOURCE
returns an exception from RIL and the rest of the response is invalid. If
last chosen user preferred network mode is CDMA based, then CDMAPhone can
have a NULL pointer exception since it accesses the fields in the
response without checking for exception. Access the fields in the
response only when exception is null.

Change-Id: Ic00e16c77cde2788a9e41e2f0a1f57c5af665d70