GPS: Changes to enable on-demand data call flow for CDMA connections
In the CDMA case we do not get a valid APN on a data connection request
to the connectivity manager. A fix is put in so that even if we get a
null APN we populate this field before we call native_agps_data_conn_open()
method to avoid a run time exception.

CRs-Fixed:231266, 243383