libstagefright: Flush audio sink when the playback is paused
When playback is paused and a reposition is done, the playback begins
with few residual frames from the paused position and then the frames
from the new position. This is because the frames which were decoded
but not rendered when playback was paused are not flushed. Flushing
those frames by calling audio sink stop.

Change-Id: Ie3e7465092f2a097b69bc820cd58b56de79bb447
CRs-Fixed: 263535