libstagefright: Flush audio sink when the playback is paused
When playback is paused and a reposition is done, the playback begins
with few residual frames from the paused position and then the frames
from the new position. This is because the frames which were decoded
but not rendered when playback was paused are not flushed. Flushing
those frames by calling audio sink stop.

Change-Id: Ib7ff085fda2b0da93b73e0a9a0b98e6a1d7d5cbf
CRs-Fixed: 263535