frameworks/base/core/java/android/widget: Refresh the progress bar on Seek completion
In RTSP streaming with Opencore, if user pauses the video followed by Suspend-Resume,
then progress goes back to start and remains there until there is any user activity on
controller window or progress bar.

In Opencore after seek player goes back to start and then seeks to the location where
user pauses before suspend. The User Interface for progress bar is created before
seek and so it shows as if playback will start from beginning of the file. If there is
no user activity then the progress bar remains at the beginning itself.

Code is added to handle seek complete indication and progress bar is updated in that.
Now even though there is no user activity progress bar is updated on seek.

Change-Id: Ic54831260f430c242e84ceda92cef993555217d6
CRs-fixed: 262120