frameworks/base/telephony: Release wakelock on RIL request send error
Android telephony does not release the partial wakelock right away if
there is an error in sending the RIL request. The wake lock is released
only after EVENT_WAKE_LOCK_TIMEOUT occurs that prevents the phone to go
in power collpase. The change is to release the wake lock as soon as the
error in send is detected.

Change-Id: I82bc44c468bf17d8905989b19731ec604e7add89