Telephony: DataServiceStateTracker to update PROPERTY_DATA_NETWORK_TYPE setting
DataServiceStateTracker tracks the data network type and so should be the one
updating the property PROPERTY_DATA_NETWORK_TYPE in the settings database. The
change is to remove the settings update from GSM and CDMA ServiceStateTracker
and add it to DataServiceStateTracker.

Also, the constant DATA_ACCESS_LTE is mapped to thirteen instead of fourteen.

Change-Id: Ia969ed0e4d58b36372f521c8b271e29d2b4ed31e