Telephony: DataServiceStateTracker to update PROPERTY_DATA_NETWORK_TYPE setting
DataServiceStateTracker tracks the data network type and so should be the one
updating the property PROPERTY_DATA_NETWORK_TYPE in the settings database. The
change is to remove the settings update from GSM and CDMA ServiceStateTracker
and add it to DataServiceStateTracker.

Also, the constant DATA_ACCESS_LTE is mapped to thirteen instead of fourteen.

Change-Id: I35f1c802f05308b4aa0bf204145a32b20814e1f9