Telephony: Initialize RIL with CDMA_SUBSCRIPTION_MODE setting
If the telephony restarts after a crash, Android RIL sets the
CDMA subscription source value from the PREFERRED_CDMA_SUBSCRIPTION
setting and that causes the mismatch with the value in the settings
database chosen by the user. The change is to use the settings value
of CDMA_SUBSCRIPTION_MODE instead when creating RIL object.

Also, remove the Setting PREFERRED_CDMA_SUBSCRIPTION from the database.

Change-Id: Ib4655ab1d58aa4d4b54d72adfd6f94808a6c20c5