Telephony: Initialize RIL with CDMA_SUBSCRIPTION_MODE setting
If the telephony restarts after a crash, Android RIL sets the
CDMA subscription source value from the PREFERRED_CDMA_SUBSCRIPTION
setting and that causes the mismatch with the value in the settings
database chosen by the user. The change is to use the settings value
of CDMA_SUBSCRIPTION_MODE instead when creating RIL object.

Change-Id: Ifffaa2ec1a4f46aaeba05d947c9f3fd243bd9640
CRs-Fixed:266034