Camera : Make the mOverlay flag NULL after calling overlay destroy in CameraService.
The Camera Service destroys the overlay in stopPreview and again before creating overlay in setOverlay.
Although the Camera Service is calling a destroy before every create correctly due to the timing at which IOverlay picks up the
destroy and create could get swapped and hence a create was followed by closing the control channel not being reflected to HAL
which goes ahead and queues or setCrops the preview frames on a closed channel.

CRs-Fixed : 267764

Change-Id: I2194e2e601e28b6c75bbd759e9247ccb1fac742b