Camera : Make the mOverlay flag NULL after calling overlay destroy in CameraService.
The Camera Service destroys the overlay in stopPreview and again before creating overlay in setOverlay.
Although the Camera Service is calling a destroy before every create correctly due to the timing at which
IOverlay picks up thedestroy and create could get swapped and hence a create was followed by closing the
control channel not being reflected to HALwhich goes ahead and queues or setCrops the preview frames on a
closed channel.

CRs-Fixed : 267764
Change-Id: I7ca6db2f0cc61b3eb309f85bfdc0fc4bd18d1866