libstagefright: Set the same input buffer size given by Qcom encoder
The function getFrameSize() does not calculate frame sizes for Tiled
formats. Qualcomm encoders already perform this calculation, and encoder
component expects the size value to be set with the same value as it
provided in getParameter().

In this change, we set the size value we get from getParameter() when using
Qcom codecs; the size will be calculated with getFrameSize() otherwise.

Change-Id: Ic1ee3cbefc238857a5b904de52594d17c4c41895