frameworks/base: acquire lock on am only when needed
appendDropBoxProcessHeaders acquires a lock on am while accessing
ProcessRecord (even if it is null). Watchdog thread ends up invoking
this function (with a null ProcessRecord) to add the stack file to
dropbox. This function would block if and when the watchdog thread is
invoked due to unavailability of lock on am resulting in a deadlock.
This would prevent watchdog from killing system_server.

(cherry picked from commit 8bdf29f888ee04caa4991deaaf484d96bab31bdd)

Change-Id: I90907570bece13de99201467a737bced714122e9