GPS: Changes to enable on-demand data call flow for CDMA connections
In the CDMA case we do not get a valid APN on a data connection request
to the connectivity manager. A fix is put in so that even if we get a
null APN we populate this field before we call native_agps_data_conn_open()
method to avoid a run time exception.

CRs-Fixed:271876, 272580
(cherry picked from commit 3c01a7561145793a134387f51383880db68f4770)

Change-Id: Icb4c2f20da20be4bc0c4d7f955f6f6a09bf66a67