Telephony: Fixes related to SIM refresh
The following issues are fixed:
1. Correct VM number updation during MBDN refresh. The voice mail number
and voice mail tag are not getting updated properly during MBDN refresh.
When the data in MBDN EF is invalid, the voice mail number and voice mail
tag are unchanged. They retain their previous value where as they should
be null in this case.

2. Handle SIM Refresh for MSISDN, CFIS, CFF_CPHS

Change-Id: Ib375e89cf9aa289a0cba3aac36d738771dcdd847
CRs-Fixed: 181461, 183830