libstagefright: Fixing a race condition while stopping video playback.
While closing the video component, two threads
hold the reference to OMXCodec - thread in
which OMXCodec::stop is executed and the thread
in which OMXCodec::on_message is executed. The
thread which holds the last reference is the
one in which OMXCodec's instance gets destroyed.
We want client thread (OMXCodec::stop) to be the
one destroying the instance. If onMessage destroys
the instance then we might hit a situation where
pthread_join is called on the same thread causing
indefinite hang.

CRs-fixed: 270937,270773
(cherry picked from commit b0c58a9d8d52021031ca6196f725282fdb0e4825)

Change-Id: I8795d812e6c58d56d055066c53b19a0d9db68610