libstagefright: Fixing a race condition while stopping video playback.
While closing the video component, two threads
hold the reference to OMXCodec - thread in
which OMXCodec::stop is executed and the thread
in which OMXCodec::on_message is executed. The
thread which holds the last reference is the
one in which OMXCodec's instance gets destroyed.
We want client thread (OMXCodec::stop) to be the
one destroying the instance. If onMessage destroys
the instance then we might hit a situation where
pthread_join is called on the same thread causing
indefinite hang.

(cherry picked from commit b0c58a9d8d52021031ca6196f725282fdb0e4825)

Change-Id: If1d0e2db2e3abe2b0847396532dcec16aae54984
CRs-fixed: 270937