frameworks: Modify getXVelocity and getYVelocity APIs of VelocityTracker class.
Some touchscreen drivers enumerate the tracking id for
the single touch from 0 whereas others enumerate it from 1.
But the getXVelocity() and getYVelocity() APIs of
VelocityTracker class only check for pointer id 0. Modify
the logic to get the pointer id from the mPointerListHead.
This is the head of a list of pointer ids sorted in
increasing order. Hence irrespective of the enumerations
by the touchscreen driver, we check for the pointer id of
the mPointerListHead node.

(cherry picked from commit 6bdaa8fab938d294924bc86dfa428b7cde8bd6d9)

Change-Id: I577ce7fcf10c0f80b400be288a52c87505d42f79
CRs-Fixed: 270706, 274403