frameworks: Modify getXVelocity and getYVelocity APIs of VelocityTracker class.
Some touchscreen drivers enumerate the tracking id for
the single touch from 0 whereas others enumerate it from 1.
But the getXVelocity() and getYVelocity() APIs of
VelocityTracker class only check for pointer id 0. Modify
the logic to get the pointer id from the mPointerListHead.
This is the head of a list of pointer ids sorted in
increasing order. Hence irrespective of the enumerations
by the touchscreen driver, we check for the pointer id of
the mPointerListHead node.

CRs-Fixed: 270706, 274403