CNE: Remove LinkProvider api dependency on default connection.
Default connection uses LinkProvider apis, remove this dependency
so that the LinkProvider apis can be disabled if necessary without
effecting the default connection.

Add the FMC property check to APIs
When FMC mode is enabled, the non-FMC CNE APIs are disabled; and
when FMC mode is disabled, the FMC APIs are disabled.

Change-Id: I0a36367f3d75572943e6dc73fce2314db5ef7513