libstagefright: Support flipping of 3D videos
If content_interpretation_type in a 3D clip's SEI is set 2 then we need
to treat the right frame as the left and the left frame as the right
(and similarly invert in case of top-bottom). This change creates a
QOMX_3D_RIGHT_LEFT_VIDEO_FLAG and QOMX_3D_BOTTOM_TOP flag for this
purpose and xors it with the color format.

Change-Id: Ifda5595a977a8908b73e145bbdfdae1070d183a5