libstagefright: Add check for server with unsupported partial GET request
When a reposition is done during HTTP streaming playback, and the requested
offset is not yet buffered, a HTTP GET request with byte range beginning
from repositioned offset to the end of clip is sent to the HTTP server.
But some HTTP servers do not support GET requests sent with the byte
range initial offset as not equal to zero. Instead the servers send data
from the beginning of the clip. This causes continuous corrupt NAL errors
and input of any key events results in an ANR. Made changes to return error
when playback encounters the above mentioned scenario.

Change-Id: Ibf02d16ca1448776e77a587381048856ca89e784