Clearing DNS Cache after connecting to another network
The code change beared in this commit resolves an issue
of a hostname resolution to different IP addresses when
connected to different networks. This issue is reported
by phone manufacturers and also is described online.

Issue Scenario Description:
Suppose the phone is browsing to a specific page (on a
hostname H) while connected to WiFi. Assume that
hostname H resolves to IP1. When WiFi is not available,
the phone gets connected to cellular data connection.
On the new network, H would resolve to IP2. The issue
is that IP1 stays in DNS cache, which does not get
flushed upon transition to the new network. The result
of IP1 being stored in DNS cache causes the browser
to try to establish connection to IP1 instead of IP2
resulting in connection failure.

The resolution delivered in this work cleans DNS Cache
for the WebView class based applications in the
following scenarios:
  WiFi SSID1 <-> WiFi SSID2
  WiFi       <-> Cellular

Change-Id: I854f5d9c0e2547076bc17f2f3eb77d1e00af4fe2