pause the video playback while waking up from SUSPEND mode
In the case when using PVPlayer as the media player, PVPlayer is
terminated when the system goes to SUSPEND while playback is on
going, and it will be loaded again and play the same video clip
once the system wakes up, since PVPlayer does not support
suspend()/resume().

The problems are 1) it plays the video from the beginning once
wakes up, not the position before suspend; 2) when the display is
on rotated LCD, the media player is unstable.

The workaround is to seek to the position before suspend and pause
the playback.

Signed-off-by: Bin Liu <b-liu@ti.com>