Telephony: CatService instance is not reinstantiated when Airplane mode is unset.
CatService is not reinstantiated on Setting and Unsetting Airplane mode.
This is because CatService instance is not being set to null when Airplane mode
is turned on. Also, the HandleThread keeps running that causes a new thread to be
created every time Airplane mode is turned off.

CRs-Fixed: 280379
(cherry picked from commit a384d9af0257062b10fa2aac44e3b09f5a680780)

Change-Id: I81ca811c6d7101975c6de1eee7c1b7eebd9d72d3