libstagefright: Don't re-kick video on seeks
Rekicking video on seek causes API incompatibilities in several
scenarios.  However the rekick was needed as the video event was
no longer active once the Video EOS occurred. This change keeps
the video event going even after a Video EOS by essentially
making it a no-op in case of video EOS.  This removes the need
for us to rekick the video event.

Change-Id: I80de711e549395e08e78bf92b3ace605eac2dcf0
CRs-Fixed: 280937