Fix DNS-cache flush when a web APK has no WiFi permission.
The fix is to always clean the DNS cache in case WebView based
Application has no ACCESS_WIFI_STATE permission when connecting
to a new WiFi network. If the persmission is in place, the network
information (SSID) is queried to assess whether connectivity to
the previously connected network is restored. If the mobile got
connected to the same network then the DNS cache does not get
flushed.

Change-Id: Ie4b17a42844821a19d4f3e051eb4f600d303966a