libstagefright: Fixing a crash in EOS for thumbnail generation.
Initializing an uninitialized variable and fixing
a crash. At EOS OMXCodec::read was returning OK and
a NULL buffer. This is not a valid return combination
as client expects buffer to be non-NULL when there
is no error. With this change OMXCodec::read
will return ERROR_END_OF_STREAM with a NULL buffer
which is a valid return combination. This change
is specific to hardware thumbnail generation.

(cherry picked from commit 4d980b627aabf2bfea50b4c7e4f265de8f2a00dd)

Change-Id: I248ffa69b2d6b1f027cb0a52b1fc5c8d9ff385e7