Telephony: CatService instance is not reinstantiated when Airplane mode is turned off
CatService is not reinstantiated on turning on and off the Airplane mode.
This is because CatService instance is not being set to null when Airplane mode
is turned on. Also, the HandleThread keeps running that causes a new thread to be
created every time Airplane mode is turned off.

Change-Id: Icb2eb1d601321d7e5750e6acd9446416edcd19c7
CRs-Fixed: 280379