libstagefright: Fix crash when generating thumbnails for AVI & h264 clips
If parser fails to identify a sample to use for thumbnail generation and
the client doesn't want thumbnail from any specific timestamp (i.e.
passes in negative timestamp), an assertion in software decoder will
fail because the software decoder attempts to start decoding from a
negative timestamp.

So instead, we'll pick up the thumbnail from timestamp 0, which we
expect to be an i-frame.

Change-Id: I2d2ccd21c7726cf06d4f3462b47106cb806cc9da
CRs-Fixed: 284873