frameworks/base: MediaRecorder: Catch exceptions when acquiring the wakelock
Recording 1080p/720p video clips on 8660 is failing due to an exception generated when
acquiring the CPU MAX wakelock while the application thread has not initialized the looper.

With this change, MediaRecorder catches the generated java exceptions when attempting
to acquire the wakelock, and then continues to record ignoring the exception.

Change-Id: I94c1144b9943de22d7860953632535f7ed3791fb
CRs-Fixed: 286413