libstagefright: Fix for race condition when stopping video playback during error state.
A crash in stagefright is happening due to a race condition
when the client calls the stop function after an error event
has come from the omx component while in executing state.

When an error event arrives during the executing state, the
stop function will move the omx component state to loaded
and wait for all messages coming from the component before
exiting.

Change-Id: Ic6aedef4d144ffd32c84d9426f91eea05c84328b
CRs-Fixed: 285154