Do not call drawWormhole in case of using overlays.
Check layerbuffercount before calling drawWormhole.
If layerbuffercount > 0, it means we are going to use overlays
so mWormholeRegion is not empty but that is fine(not a wm bug).
In that case we call drawWormhole which issues some gl commands
without calling glfinish or swapbuffers(because we use overlays).
That causes mem leak as the gl driver stack keeps on growing.
CRs-fixed: 280056

Change-Id: Id788175581f64046c09c8b0db758d920b18a1d89