Do not call drawWormhole in case of using overlays.
Check layerbuffercount before calling drawWormhole.
If layerbuffercount > 0, it means we are going to use overlays
so mWormholeRegion is not empty but that is fine(not a wm bug).
In that case we call drawWormhole which issues some gl commands
without calling glfinish or swapbuffers(because we use overlays).
That causes mem leak as the gl driver stack keeps on growing.
CRs-fixed: 280056

(cherry picked from commit 9b79db8f345480c635b5d354d1437a479cda04ea)

Change-Id: I438732d0242568e5b7ab925af0e8c8fafba2730b