frameworks/base: Increase the buffer size to hold the whole content of the file.
When we have more cpufreq steps and cpufreq's time_in_state stats file
can grow beyond 256.

This buffer size was not enough because the file has more than 300
characters. So it read part of the file and the parsing failed.
This runs from ActivityManager, so when the ActivityManager dies
framework reboots.Increasing the buffer size will help use to get
the whole file into the buffer.

Change-Id: I80dd7e1ff790707434f563a7e57f7e162fab488b