Telephony: Block multiple Call switch request to RIL
Sending a call switch request to RIL before receiving
a response for the previous request causes the lower
layers to not respond. Pending responses from RIL
block the call polling in telephony.

So blocked sending a call switch request to ril
when there is a response pending already.
This way even if the user hits the hold/unhold
button continuously telephony will reject the
multiple requests from reaching RIL.

Change-Id: I353504fc3c3d3faa9332a59d5a8713f511351f98
CRs-Fixed: 282061