SurfaceFlinger: Full screen clip for push buffers
When using MDP composition, there is a
rounding issue in copybit where
stretching part of an image of smaller
dimensions to a larger area causes a
distortion of a few pixels. This is most
apparent when playing video or in camera
preview. Setting the clip for push buffers
to the hardware bounds fixes the issue

CRs-fixed: 282961
(cherry picked from commit e20d309004132cb41a8bd265274c71ca5da22bdd)

Change-Id: Ieb8feef980a9b01eddba6aeb6b34434e3b85bca8