SurfaceFlinger: Full screen clip for push buffers
When using MDP composition, there is a
rounding issue in copybit where
stretching part of an image of smaller
dimensions to a larger area causes a
distortion of a few pixels. This is most
apparent when playing video or in camera
preview. Setting the clip for push buffers
to the hardware bounds fixes the issue

CRs-fixed: 282961
(cherry picked from commit e20d309004132cb41a8bd265274c71ca5da22bdd)

Change-Id: I30a5ff7f60cb8d4e4a2126795c6571844a947c3e