SurfaceFlinger: Full screen clip for push buffers
When using MDP composition, there is a
rounding issue in copybit where
stretching part of an image of smaller
dimensions to a larger area causes a
distortion of a few pixels. This is most
apparent when playing video or in camera
preview. Setting the clip for push buffers
to the hardware bounds fixes the issue

Change-Id: I2b9dd3975019197278b9b3427ca18157b3cf1a5c
CRs-fixed: 282961