OMXCodec changes for tearing effect
In the OMXCodec buffer info, a new field has been added to track the
buffers which are with Player. This means the buffer could be in AV
sync logic, or already with DSS for rendering in case of video buffers,
or with Audio HAL incase of Audio buffers. Earlier these buffers are
given back to codec for filling irrespective of whether they are
with sink or free.

This is fixed by adding additional state for the buffer to track the
life cycle of the buffer with the sink.

This fixes the tearing video frame issue on display when
fastford/rewind during the video playback.

Signed-off-by: Bin Liu <b-liu@ti.com>