TSLIB: Added support for multiple event reads from driver (at max 5).
Modified getEvent function to process the TS events only when
tsSampReady=1 i.e. when a SYN event has been received from the
driver.

This corrects the problem where in some of the cases, ts_read was
reporting previous (x,y,p) co-ordinates along with current
co-ordinates. This was causing a line to be drawn when a point was
expected.

Change-Id: Ib24c6e51f9030201cf664917066db37276723552
CRs-fixed: 283825