Update declared OMX version to 1.1.0
This change resolves a OMX_ErrorVersionMismatch error from having a mismatch
between the OMX versions in the StageFright data structures and those reported
by the OpenMax components being used.  Everything is using OpenMax 1.1.2, but
StageFright is declaring 1.0.0 and the components are declaring 1.1.0.  This
change makes them align on 1.1.0, but ultimately both should be set to 1.1.2.