StageFright TestApp:Video dec frame rate support
Integration of this patch is as a part of am45x bring up on rowboat

Patch Source link#
git://git.omapzoom.org/platform/frameworks/base.git

History of this patch is as follows:

commit dee20efc8d18a0a2cb754b44e36f309aa6d0759d
Author: a0393586 <a0393586@ti.com>
Date:   Tue Feb 15 18:16:56 2011 +0530

StageFright Test App: Video decoding frame rate support

The changes are for porting the below patch from Froyo to Gingerbread
----------------------------------------------------------
http://review.omapzoom.org/#change,6409

Currently the stagefright test app does the video decoding continuously
without depending on the frame rate.

Have modified the test app to handle the decoding based on frame rate.
It can be passed as input parameter in command line and the clips would
be decoded accordingly.
Eg: "-f 30" would ensure that clip is decoded at 30fps.

Incase this option is not specified it would decode continuously as
earlier.
-----------------------------------------------------------
Author:dnk.mohan@ti.com
Signed-off-by: satish <satish.patel@ti.com>