frameworks/base: Copy bypass buffer when display is frozen.
When comp. bypass is in use and display is frozen,
application might need to go ahead. This will fail since,
current buffer is marked as use by SF including bypass buffer.
And, it causes hang. To make application go ahead, we copy the
current buffer to rotator buffer when display is frozen.

Change-Id: I1afd22215984581032da13ded909f07842573058
CRs-fixed: 288992