SurfaceFlinger: Invalidate screen when playing video
There is a rounding issue in copybit where
stretching part of an image of smaller
dimensions to a larger area causes a
distortion of a few pixels. This is most
apparent when the clip is smaller than
the transformed region of the video.
Hence, when using MDP composition,
invalidating the region to the transformed
region of the video.

Change-Id: I0974cf75ea5be8ec921a0ccd19ec83925d03dabf
CRs-fixed: 282961