SurfaceFlinger: Invalidate screen when playing video
There is a rounding issue in copybit where
stretching part of an image of smaller
dimensions to a larger area causes a
distortion of a few pixels. This is most
apparent when the clip is smaller than
the transformed region of the video.
Hence, when using MDP composition,
invalidating the region to the transformed
region of the video.

CRs-fixed: 282961
(cherry picked from commit a57c8109fc52cebe234d929d262dd7fe19ea5994)

Change-Id: I0d4e9630dc053cf7335e49633be8dcb6a3256a40