Telephony: Trigger NV ready event to load ERI file.
When phone is in NV mode from bootup, event NV ready is not triggered
currently in CDMA Phone. This is because the default cdma subscription
source assumed in CDMA Phone object is NV and the RIL reported
cdma subscription source is also nv. NV ready is posted only when
there is a change in subscription source from RUIM to NV. Changes
are made so that NV ready is triggered always if it is an NV subscription.

Change-Id: I7864bb881bcd20ac9078a8b32cac7a48594ecae8
CRs-Fixed: 290276