frameworks/base: Show airplane mode icon only when radio is off
If user turns the airplane mode the airplane mode icon is displayed
even if the modem is unavailable (STATE_OUT_OF_SERVICE) as StatusBarPolicy
allows airplane mode icon to displayed for both out of service and radio
off states. If the modem is not available then the RADIO_POWER request is
not even sent to the modem so doesn't make sense to show the airplane mode
icon. Also, if the modem is available but didn't respond to the RADIO_POWER
then there is no comfirmation that the radio is off so we shouldn't display
the airplane mode icon. The fix is to display the airplane mode icon only if
the radio state is off.

Change-Id: Id4cf405303dbbb82673d2a36e78641760365416b
CRs-Fixed: 291235