frameworks/base: Show airplane mode icon only when radio is off
If user turns the airplane mode the airplane mode icon is displayed
even if the modem is unavailable (STATE_OUT_OF_SERVICE) as StatusBarPolicy
allows airplane mode icon to displayed for both out of service and radio
off states. If the modem is not available then the RADIO_POWER request is
not even sent to the modem so doesn't make sense to show the airplane mode
icon. Also, if the modem is available but didn't respond to the RADIO_POWER
then there is no comfirmation that the radio is off so we shouldn't display
the airplane mode icon. The fix is to display the airplane mode icon only if
the radio state is off.

CRs-Fixed: 291235
(cherry picked from commit 1b5a33f7da1c89367923df2b0da32c3db24b8a8d)

Change-Id: I3c67e5161ba930b2e3048ef2a08e0433f7d742a9