frameworks/base: bluetooth: Fix to check the SRM header status in a Put/Get Operation
OBEX client was initiating a SRM mode of operation based on the SRM status returned
in the Connect response from OBEX server. As per specification of GOEPv2 it is not
mandatory to send the SRM_SUPPORTED in the connect response. As a side effect the SRM
mode was not getting enabled for OBEX operation with OBEX Server not sending the
SRM_SUPPORTED in connect response.

The change adds support to handle the decision to enable SRM  based on the Response
for Put/Get Operation.

Change-Id: I4956402160ac8778e3b632f3de892b05ddccccad